feat: spot db

+++ pkg/bn/utils/bn_clientId.go
@@ -6,7 +6,7 @@ import (
 )
 
 // default client id with running number from symbol
-func BinanceDefaultClientID(symbol string, position_side string, counting int) string {
+func BinanceDefaultClientID(symbol string, position_side string, counting int64) string {
 	return fmt.Sprintf("%s_%s_%d", symbol, position_side, counting)
 }
